refactor: modularize backend server
