+++ packages/backend/src/utils/connections.js
@@ -1,0 +1,3 @@
+const connections = new Map();
+
+module.exports = connections;
